+++ CardPayments/src/main/java/com/paypal/android/cardpayments/CardVaultRequest.kt
@@ -10,6 +10,7 @@ import kotlinx.parcelize.Parcelize
  *
  * @property setupTokenId id for the setup token to update.
  * @property card card payment source to attach to the setup token.
+ * @property returnUrl return url for deep linking back into the merchant app after an auth challenge
  */
 @Parcelize
 data class CardVaultRequest(