+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 * FraudDetection
   * Bump Magnes dependency to version 5.4.0
 * PaymentButtons
-  * Font typeface changed to "PayPalOpen" to meet brand guidelines
+  * Update font typeface to "PayPalOpen" to meet brand guidelines
 
 ## 1.3.0 (2024-01-09)
 * PaymentButtons