+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## unreleased
 * PaymentButtons
+  * Deprecate `PayPalCreditButtonColor.GOLD` and `.WHITE`
   * Undeprecate `PayPalCreditButtonColor.BLACK` and `.DARK_BLUE`
   * Undeprecate `PayPalButtonColor.BLUE`, `.BLACK`, and `.SILVER`
 * CardPayments