+++ CHANGELOG.md
@@ -5,7 +5,7 @@
   * Undeprecate `PayPalCreditButtonColor.BLACK` and `.DARK_BLUE`
   * Undeprecate `PayPalButtonColor.BLUE`, `.BLACK`, and `.SILVER`
   * Added analytics events
-    * `paypal-button:initialized` and `paypal-button:tapped`
+    * `payment-button:initialized` and `payment-button:tapped`
 * CardPayments
   * Add `liabilityShift` property to `CardResult`
   * Callback `PayPalSDKError` when `CardClient#approveOrder()` 3DS verification fails